Added a centre of mass method, began building an "Annotation" object
The Annotation object is intended to encapsulate the whole of an annotation, but will probably be generated as a collection from an image of annotations using a pretty simple "get a collection of contiguous areas" method.

I also want to use a fancy-pants binary search tree to assign annotations to the appropriate part of the training array:

http://www.cs.princeton.edu/courses/archive/spring13/cos226/lectures/99GeometricSearch.pdf

probably going to use PIL for image inhalation.