This should do all the segmentation of Annotations
So I've created another class, AnnotationFactory, that takes in an annotation image (that should ideally be a series of contiguous cell annotations that are not connected together) and traverses it looking for hot pixels. It marks each pixel pretty much as soon as it arrives so we don't get loops.

When it finds a hot pixel, it then makes an Annotation object and searches that pixel's neighbours (' neighbours' neighbours), passing the Annotation in. If a neighbour has been visited, the thing immediately kicks out. If the neighbour is not hot, it marks the neighbour visited but doesn't search any of THAT neighbour's neighbours. If the neighbour is hot, it adds it to the annotation passed in and continues.

In this way, cold pixels should be visited by
- traversal
- being adjacent to an annotation

and hot pixels should be visited by
- traversal
- and then every contiguous hot pixel gets picked up by the recursion.

I have no idea how deep the stack in Python goes but if the cells are under a few hundred pixels wide this shouldn't bother us.

In addition, I have now wired up the centre of mass method so once an Annotation has been made, it gets the COM.

TODO: convert the array of hot and cold pixels in an annotation (which is currently the size of the whole image) to an m by m patch.