Added the ability to measure and invalidate a bounding box
Bbox done by the usual "is this hot pixel outside our minima and maxima" thing, but if we set a COLD pixel then a flag is set to invalidate the bbox. Then if you call the bbox retrieve method it does a JIT recalculate. I don't anticipate setting a pixel as cold explicitly but it's nice to be covered!