The TF model compiles!
There's still a weird relationship between filters and kernels that I don't understand. Specifically, it seems to require that filters on the previous layer is a bigger number than kernels on the current one, which is bizarre because I just want to scan a single image for a bunch of layers, and apparently that is Not A Thing.

It may be that it's expanding things to tuples in a way I don't follow.

Anyway: we march onwards